Resolvi el tema de los enums para las distintas formas de pago. Lo testeé en propietario , solo faltó hacerlo en publicacion para cada caso de si recibe todos los medios de pago o alguno.

+++ src/unq/Puntaje.java
@@ -26,7 +26,7 @@ public class Puntaje {
 		this.categoria = categoria;
 	}
 
-	public int getValor() {
+	public double getValor() {
 		return (this.valor);
 	}
